frontend_base/requirements.txt
streamlit>=1.37.0
pandas>=2.2.2
plotly>=5.23.0
firebase-admin>=6.5.0
google-cloud-firestore>=2.17.0
